feat: change rpm from int to float - change all references for rpm from int to float - update calculation for interval step time in Stepper

+++ lib/Stepper/Stepper.h
@@ -9,13 +9,13 @@ public:
     void set_timer(int timer_number, void (*isr)());
 
     void setup_pins();
-    void set_rpm(int _rpm);
+    void set_rpm(float _rpm);
     int step_once();
     //void update_stepper();
     void manual_steps(int steps);
     int get_steps_remaining(){return this->steps_remaining;};
     int get_steps_per_rev(){return this->steps_per_rev;};
-    unsigned int get_step_interval_us(){return this->step_interval_us;};
+    uint32_t get_step_interval_us(){return this->step_interval_us;};
     
 private:
     int motor_pin1; //blue
@@ -23,9 +23,9 @@ private:
     int motor_pin3; //yellow
     int motor_pin4; //orange
     int steps_per_rev;
-    int speed_rpm;
+    float speed_rpm;
     volatile int steps_remaining; // will need to sort this out for continuous mode, probably just put an if statement in step_once
-    unsigned int step_interval_us;
+    uint32_t step_interval_us;
     volatile int current_state;
     int direction; // ccw = 1, cw = -1 (ccw is given in datasheet)
     hw_timer_t *Timer_cfg;
